Silencing compiler warnings - round 2: vanilla non-windowport
Fixes warnings in include/ and src/ and also util/dlb_main.c. Most of
these are implicit fallthroughs; I added break statements rather than
make the fallthru explicit because I doubt the behavior was intended
that way. This has minor gameplay ramifications: tools sort slightly
differently, and healing potions no longer act like a dipped unicorn
horn does on bad potions.

Also: force a couple things to be unsigned to squash sign-compare and
char-subscripts warnings.

Many warnings still exist in windowport code. This commit does not
address any of those.

+++ include/hack.h
@@ -208,7 +208,7 @@ typedef struct sortloot_item Loot;
     (Warn_of_mon && ((context.warntype.obj                                   \
                       && (((context.warntype.obj & 0x80000000) != 0) ?       \
                           ((context.warntype.obj & 0x7FFFFFFF)               \
-                                                     == (mon)->data->mlet) : \
+                                      == (unsigned int) (mon)->data->mlet) : \
                           (context.warntype.obj & (mon)->data->mflags2)))    \
                      || (context.warntype.polyd                              \
                          && (context.warntype.polyd & (mon)->data->mflags2)) \
